+++ Docs/Changes.md
@@ -1,6 +1,7 @@
 # 0.25.0
   * Add: _Add Integer.isMultiple(of) method_ ([Issue #454](https://github.com/vorov2/dyalect/issues/454), [docs](https://github.com/vorov2/dyalect/wiki/Integer#isMultiple)).
   * Add: _Map: add method `compact(by)`_ ([Issue #393](https://github.com/vorov2/dyalect/issues/393), [docs](https://github.com/vorov2/dyalect/wiki/Map#compact)).
+  * Add: _Add `Integer.parse` method_ ([Issue #480](https://github.com/vorov2/dyalect/issues/480), [docs](https://github.com/vorov2/dyalect/wiki/Integer#parse)).
 
 # 0.24.0
   * Fix: _`Map` doesn't support `iter`_ ([Issue #473](https://github.com/vorov2/dyalect/issues/473))